+++ src/devices/DevicesService.ts
@@ -3,18 +3,18 @@ import { ObjectId } from 'mongodb'
 
 export const devicesService = {
   async getDevicesByUserId(userId: string) {
-    return await devicesCollection.find({ userId }).toArray()
+    return await devicesCollection.find({ userId: new ObjectId(userId) }).toArray()
   },
 
   async deleteAllDevicesExceptCurrent(userId: string, currentRefreshToken: string) {
     const currentDevice = await devicesCollection.findOne({
-      userId,
+      userId: new ObjectId(userId),
       refreshToken: currentRefreshToken,
     })
     if (!currentDevice) return false
 
     await devicesCollection.deleteMany({
-      userId,
+      userId: new ObjectId(userId),
       _id: { $ne: new ObjectId(currentDevice._id) },
     })
 
@@ -22,7 +22,7 @@ export const devicesService = {
   },
 
   async deleteDeviceById(userId: string, deviceId: string) {
-    const result = await devicesCollection.deleteOne({ userId, deviceId })
+    const result = await devicesCollection.deleteOne({ userId: new ObjectId(userId), deviceId })
     return result.deletedCount > 0
   },
 }
